newversion v7(final)/EdgeSolution/modules/analyticmodule/requirements.txt
azure-iot-device
pandas
schedule